+++ src/projectgoldstarscalculators/Divide.java
@@ -5,11 +5,11 @@ import java.awt.event.ActionListener;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JTextField;
-public class CalcDivide implements ActionListener
+public class Divide implements ActionListener
 {
     public static JTextField tf1, tf2, tf3, tf4, tf5, tf6, tf7, tf8, tf9, tf10;
     
-    public CalcDivide()
+    public Divide()
     {
         divide();
     }